requirements.txt
autopep8==1.5.5
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==2.2.4
idna==2.10
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
